requirements.txt
# Web scraping and browser automation
playwright==1.48.0
requests==2.32.3

# Database
supabase==2.10.0
psycopg2-binary==2.9.10

# AI/ML for embeddings
torch==2.4.0
torchvision==0.19.0
transformers==4.40.0
Pillow==10.3.0

# Configuration and utilities
PyYAML==6.0.2
python-dotenv==1.0.1
loguru==0.7.2

# Data processing
pandas==2.2.3
numpy==1.26.4

# Development
pytest==8.3.3
black==24.10.0
flake8==7.1.1
